Updated Authentication and Authorization With Cookies
I created the "auth_cookie_function" file.
This file implement authentication and authorization using cookies.
The "bootstrap" file was updated with a link to the  "auth_cookie_function" file.
The "auth_function" file was disabled.
The "do_logout" function has been updated to use the "set_userinfo_cookie" to logout setting the time to a time in the past